+++ app/assets/javascripts/mobile/mobile_comments.js
@@ -7,6 +7,8 @@
 (function() {
   Diaspora.Mobile = {};
   Diaspora.Mobile.Comments = {
+    stream: function(){ return $(".stream"); },
+
     initialize: function() {
       var self = this;
       $(".stream").on("tap click", "a.back_to_stream_element_top", function() {
@@ -17,96 +19,135 @@
         }, 1000);
       });
 
-      $(".stream").on("tap click", "a.show_comments", function(evt){
+      this.stream().on("tap click", "a.show-comments", function(evt){
         evt.preventDefault();
         self.toggleComments($(this));
       });
 
-      $(".stream").on("tap click", "a.comment-action", function(evt) {
+      this.stream().on("tap click", "a.comment-action", function(evt) {
         evt.preventDefault();
         self.showCommentBox($(this));
-        var bottomBar = $(this).closest(".bottom_bar").first();
-        var commentContainer = bottomBar.find(".comment_container").first();
+        var bottomBar = $(this).closest(".bottom-bar").first();
+        var commentContainer = bottomBar.find(".comment-container").first();
         self.scrollToOffset(commentContainer);
       });
 
-      $(".stream").on("submit", ".new_comment", this.submitComment);
+      this.stream().on("submit", ".new_comment", this.submitComment);
     },
 
-    submitComment: function(evt) {
+    submitComment: function(evt){
       evt.preventDefault();
       var form = $(this);
       var commentBox = form.find(".comment_box");
       var commentText = $.trim(commentBox.val());
-      if (commentText) {
-        $.post(form.attr("action")+"?format=mobile", form.serialize(), function(data) {
-          Diaspora.Mobile.Comments.updateStream(form, data);
-        }, "html");
-        return true;
-      }
-      else {
+      if(!commentText){
         commentBox.focus();
         return false;
       }
+
+      $.post(form.attr("action") + "?format=mobile", form.serialize(), function(data){
+        Diaspora.Mobile.Comments.updateStream(form, data);
+      }, "html").fail(function(){
+        Diaspora.Mobile.Comments.resetCommentBox(this);
+      });
+
+      autosize($(".add-comment-switcher:not(.hidden) textarea"));
     },
 
     toggleComments: function(toggleReactionsLink) {
       if(toggleReactionsLink.hasClass("loading")) { return; }
-      if (toggleReactionsLink.hasClass("active")) {
+
+      if(toggleReactionsLink.hasClass("active")) {
         this.hideComments(toggleReactionsLink);
+        toggleReactionsLink.parents(".bottom-bar").find(".add-comment-switcher").addClass("hidden");
       } else {
         this.showComments(toggleReactionsLink);
       }
     },
 
     hideComments: function(toggleReactionsLink) {
-      var bottomBar = toggleReactionsLink.closest(".bottom_bar").first(),
-          commentsContainer = this.commentsContainerLazy(bottomBar),
-          existingCommentsContainer = commentsContainer();
-      existingCommentsContainer.hide();
+      var bottomBar = toggleReactionsLink.closest(".bottom-bar").first();
+      this.bottomBarLazy(bottomBar).deactivate();
       toggleReactionsLink.removeClass("active");
     },
 
     showComments: function(toggleReactionsLink) {
-      var bottomBar = toggleReactionsLink.closest(".bottom_bar").first(),
-          commentsContainer = this.commentsContainerLazy(bottomBar),
-          existingCommentsContainer = commentsContainer(),
+      var bottomBar = toggleReactionsLink.closest(".bottom-bar").first(),
+          bottomBarContainer = this.bottomBarLazy(bottomBar),
+          existingCommentsContainer = bottomBarContainer.getCommentsContainer(),
           commentActionLink = bottomBar.find("a.comment-action");
+
+      bottomBarContainer.activate();
+      bottomBarContainer.showLoader();
+
       if (existingCommentsContainer.length > 0) {
         this.showLoadedComments(toggleReactionsLink, existingCommentsContainer, commentActionLink);
+        bottomBarContainer.hideLoader();
       } else {
         this.showUnloadedComments(toggleReactionsLink, bottomBar, commentActionLink);
       }
     },
 
     showLoadedComments: function(toggleReactionsLink, existingCommentsContainer, commentActionLink) {
-      toggleReactionsLink.addClass("active");
-      existingCommentsContainer.show();
       this.showCommentBox(commentActionLink);
       existingCommentsContainer.find("time.timeago").timeago();
     },
 
     showUnloadedComments: function(toggleReactionsLink, bottomBar, commentActionLink) {
       toggleReactionsLink.addClass("loading");
-      var commentsContainer = this.commentsContainerLazy(bottomBar);
+      var bottomBarContainer = this.bottomBarLazy(bottomBar);
       var self = this;
       $.ajax({
         url: toggleReactionsLink.attr("href"),
         success: function (data) {
           toggleReactionsLink.addClass("active").removeClass("loading");
-          $(data).insertAfter(bottomBar.children(".show_comments").first());
+          $(data).insertAfter(bottomBar.children(".show-comments").first());
           self.showCommentBox(commentActionLink);
-          commentsContainer().find("time.timeago").timeago();
+          bottomBarContainer.getCommentsContainer().find("time.timeago").timeago();
+          bottomBarContainer.activate();
         },
-        error: function() {
-          toggleReactionsLink.removeClass("loading");
+        error: function(){
+          bottomBarContainer.deactivate();
         }
+      }).always(function(){
+        toggleReactionsLink.removeClass("loading");
+        bottomBarContainer.hideLoader();
       });
     },
 
-    commentsContainerLazy: function(bottomBar) {
-      return function() {
-        return bottomBar.find(".comment_container").first();
+    bottomBarLazy: function(bottomBar) {
+      return  {
+        loader: function(){
+          return bottomBar.find(".ajax-loader");
+        },
+
+        getCommentsContainer: function(){
+          return bottomBar.find(".comment-container").first();
+        },
+
+        getShowCommentsLink: function(){
+          return bottomBar.find("a.show-comments");
+        },
+
+        showLoader: function(){
+          this.loader().removeClass("hidden");
+        },
+
+        hideLoader: function(){
+          this.loader().addClass("hidden");
+        },
+
+        activate: function(){
+          bottomBar.addClass("active").removeClass("inactive");
+          this.getShowCommentsLink().addClass("active");
+          this.getShowCommentsLink().find("i").removeClass("entypo-chevron-down").addClass("entypo-chevron-up");
+        },
+
+        deactivate: function(){
+          bottomBar.removeClass("active").addClass("inactive");
+          this.getShowCommentsLink().removeClass("active");
+          this.getShowCommentsLink().find("i").addClass("entypo-chevron-down").removeClass("entypo-chevron-up");
+        }
       };
     },
 
@@ -121,72 +162,76 @@
     },
 
     showCommentBox: function(link){
-      if(!link.hasClass("inactive") || link.hasClass("loading")) { return; }
-      var self = this;
-      $.ajax({
-        url: link.attr("href"),
-        beforeSend: function(){
-          link.addClass("loading");
-        },
-        context: link,
-        success: function(data) {
-          self.appendCommentBox.call(this, link, data);
-        },
-        error: function() {
-          link.removeClass("loading");
-        }
-      });
-    },
-
-    appendCommentBox: function(link, data) {
-      link.removeClass("loading");
-      link.removeClass("inactive");
-      var bottomBar = link.closest(".bottom_bar").first();
-      bottomBar.append(data);
+      var bottomBar = link.closest(".bottom-bar").first();
       var textArea = bottomBar.find("textarea.comment_box").first()[0];
+      bottomBar.find(".add-comment-switcher").removeClass("hidden");
       autosize(textArea);
     },
 
     updateStream: function(form, data) {
-      var bottomBar = form.closest(".bottom_bar").first();
+      var bottomBar = form.closest(".bottom-bar").first();
       this.addNewComments(bottomBar, data);
       this.updateCommentCount(bottomBar);
-      this.updateReactionCount(bottomBar);
+      this.increaseReactionCount(bottomBar);
       this.handleCommentShowing(form, bottomBar);
       bottomBar.find("time.timeago").timeago();
     },
 
     addNewComments: function(bottomBar, data) {
-      if ($(".comment_container", bottomBar).length === 0) {
-        $(".show_comments", bottomBar).after($("<div/>", {"class": "comment_container"}));
-        $(".comment_container", bottomBar).append($("<ul/>", {"class": "comments"}));
+      if ($(".comment-container", bottomBar).length === 0) {
+        $(".show-comments", bottomBar).after($("<div/>", {"class": "comment-container"}));
+        $(".comment-container", bottomBar).append($("<ul/>", {"class": "comments"}));
       }
-      $(".comment_container .comments", bottomBar).append(data);
+      $(".comment-container .comments", bottomBar).append(data);
     },
 
     // Fix for no comments
     updateCommentCount: function(bottomBar) {
-      var commentCount = bottomBar.find(".comment_count");
+      var commentCount = bottomBar.find(".comment-count");
       commentCount.text(commentCount.text().replace(/(\d+)/, function (match) {
-        return parseInt(match) + 1;
+        return parseInt(match, 10) + 1;
       }));
     },
 
     // Fix for no reactions
-    updateReactionCount: function(bottomBar) {
-      var toggleReactionsLink = bottomBar.find(".show_comments").first();
-      toggleReactionsLink.text(toggleReactionsLink.text().replace(/(\d+)/, function (match) {
-        return parseInt(match) + 1;
-      }));
+    increaseReactionCount: function(bottomBar) {
+      var toggleReactionsLink = bottomBar.find(".show-comments").first();
+      var count = toggleReactionsLink.text().match(/.*(\d+).*/);
+      var text = "";
+
+      // No previous comment
+      if(!count){
+        text = Diaspora.I18n.t("stream.reactions", {count: 1});
+        var parent = toggleReactionsLink.parent();
+        var postGuid = bottomBar.parents(".stream_element").data("guid");
+
+        toggleReactionsLink.remove();
+        toggleReactionsLink = $("<a/>", {"class": "show-comments", "href": Routes.postComments(postGuid) + ".mobile"})
+          .html(text + "<i class='entypo-chevron-up'/>");
+        parent.prepend(toggleReactionsLink);
+        bottomBar.removeClass("inactive").addClass("active");
+      }
+      else {
+        count = parseInt(count, 10) + 1;
+        text = Diaspora.I18n.t("stream.reactions", {count: count});
+        toggleReactionsLink.html(text + "<i class='entypo-chevron-up'/>");
+      }
     },
 
     handleCommentShowing: function(form, bottomBar) {
       var formContainer = form.parent();
-      formContainer.remove();
+      formContainer.find("textarea.form-control").first().val("");
+      this.resetCommentBox(formContainer);
       var commentActionLink = bottomBar.find("a.comment-action").first();
       commentActionLink.addClass("inactive");
-      var toggleReactionsLink = bottomBar.find(".show_comments").first();
-      this.showComments(toggleReactionsLink);
+      this.showComments(bottomBar.find(".show-comments").first());
+    },
+
+    resetCommentBox: function(el){
+      var commentButton = el.find("input.comment-button").first();
+      commentButton.attr("value", commentButton.data("reset-with"));
+      commentButton.removeAttr("disabled");
+      commentButton.blur();
     }
   };
 })();