+++ spec/javascripts/mobile/mobile_comments_spec.js
@@ -2,7 +2,7 @@ describe("Diaspora.Mobile.Comments", function(){
   describe("toggleComments", function() {
     beforeEach(function() {
       spec.loadFixture("aspects_index_mobile_post_with_comments");
-      this.link = $(".stream .show_comments").first();
+      this.link = $(".stream .show-comments").first();
       spyOn(Diaspora.Mobile.Comments, "showComments");
       spyOn(Diaspora.Mobile.Comments, "hideComments");
     });
@@ -31,27 +31,27 @@ describe("Diaspora.Mobile.Comments", function(){
   describe("showUnloadedComments", function() {
     beforeEach(function() {
       spec.loadFixture("aspects_index_mobile_post_with_comments");
-      this.link = $(".stream .show_comments").first();
-      this.bottomBar = this.link.closest(".bottom_bar").first();
+      this.link = $(".stream .show-comments").first();
+      this.bottomBar = this.link.closest(".bottom-bar").first();
       this.commentActionLink = this.bottomBar.find("a.comment-action");
     });
 
     it("adds the 'loading' class to the link", function() {
       Diaspora.Mobile.Comments.showUnloadedComments(this.link, this.bottomBar, this.commentActionLink);
-      expect($(".show_comments").first()).toHaveClass("loading");
+      expect($(".show-comments").first()).toHaveClass("loading");
     });
 
     it("removes the 'loading' class if the request failed", function() {
       Diaspora.Mobile.Comments.showUnloadedComments(this.link, this.bottomBar, this.commentActionLink);
       jasmine.Ajax.requests.mostRecent().respondWith({status: 400});
-      expect($(".show_comments").first()).not.toHaveClass("loading");
+      expect($(".show-comments").first()).not.toHaveClass("loading");
     });
 
     it("adds the 'active' class if the request succeeded", function() {
       Diaspora.Mobile.Comments.showUnloadedComments(this.link, this.bottomBar, this.commentActionLink);
       jasmine.Ajax.requests.mostRecent().respondWith({status: 200, contentType: "text/plain", responseText: "test"});
-      expect($(".show_comments").first()).toHaveClass("active");
-      expect($(".show_comments").first()).not.toHaveClass("loading");
+      expect($(".show-comments").first()).toHaveClass("active");
+      expect($(".show-comments").first()).not.toHaveClass("loading");
     });
 
     it("calls showCommentBox", function() {
@@ -70,50 +70,12 @@ describe("Diaspora.Mobile.Comments", function(){
       });
       expect($(".stream .stream_element").first()).toContainElement(".commentContainerForTest");
     });
-  });
 
-  describe("showCommentBox", function() {
-    beforeEach(function() {
-      spec.loadFixture("aspects_index_mobile_post_with_comments");
-      this.link = $(".stream .comment-action").first();
-    });
-
-    it("adds the 'loading' class to the link", function() {
-      Diaspora.Mobile.Comments.showCommentBox(this.link);
-      expect($(".comment-action").first()).toHaveClass("loading");
-    });
-
-    it("removes the 'loading' class if the request failed", function() {
-      Diaspora.Mobile.Comments.showCommentBox(this.link);
-      jasmine.Ajax.requests.mostRecent().respondWith({status: 400});
-      expect($(".comment-action").first()).not.toHaveClass("loading");
-    });
-
-    it("fires an AJAX call", function() {
-      spyOn(jQuery, "ajax");
-      Diaspora.Mobile.Comments.showCommentBox(this.link);
-      expect(jQuery.ajax).toHaveBeenCalled();
-    });
-
-    it("calls appendCommentBox", function() {
-      spyOn(Diaspora.Mobile.Comments, "appendCommentBox");
-      Diaspora.Mobile.Comments.showCommentBox(this.link);
+    it("shows and hides the mobile spinner", function(){
+      Diaspora.Mobile.Comments.showComments(this.link);
+      expect($(".ajax-loader").first()).toBeVisible();
       jasmine.Ajax.requests.mostRecent().respondWith({status: 200, contentType: "text/plain", responseText: "test"});
-      expect(Diaspora.Mobile.Comments.appendCommentBox).toHaveBeenCalledWith(this.link, "test");
-    });
-
-    it("doesn't do anything if the link class is 'loading'", function() {
-      spyOn(jQuery, "ajax");
-      this.link.addClass("loading");
-      Diaspora.Mobile.Comments.showCommentBox(this.link);
-      expect(jQuery.ajax).not.toHaveBeenCalled();
-    });
-
-    it("doesn't do anything if the link class is not 'inactive'", function() {
-      spyOn(jQuery, "ajax");
-      this.link.removeClass("inactive");
-      Diaspora.Mobile.Comments.showCommentBox(this.link);
-      expect(jQuery.ajax).not.toHaveBeenCalled();
+      expect($(".ajax-loader").first()).not.toBeVisible();
     });
   });
 
@@ -136,6 +98,73 @@ describe("Diaspora.Mobile.Comments", function(){
       spyOn(jQuery, "ajax");
       form.submit();
       expect(jQuery.ajax).not.toHaveBeenCalled();
+    });
+  });
+
+  describe("increaseReactionCount", function(){
+    beforeEach(function() {
+      spec.loadFixture("aspects_index_mobile_post_with_comments");
+      this.bottomBar = $(".bottom-bar").first();
+      this.toggleReactionsLink = this.bottomBar.find(".show-comments").first();
+    });
+
+    it("Increase reaction count from 1", function(){
+      expect(this.toggleReactionsLink.text().trim()).toBe("5 reactions");
+      Diaspora.Mobile.Comments.increaseReactionCount(this.bottomBar);
+      expect(this.toggleReactionsLink.text().trim()).toBe("6 reactions");
+    });
+
+    it("Creates the reaction link when no reactions", function(){
+      var parent = this.toggleReactionsLink.parent();
+      var postGuid = this.bottomBar.parents(".stream_element").data("guid");
+      this.toggleReactionsLink.remove();
+      parent.prepend($("<span/>", {"class": "show-comments"}).text("No reaction"));
+
+      Diaspora.Mobile.Comments.increaseReactionCount(this.bottomBar);
+      this.toggleReactionsLink = this.bottomBar.find(".show-comments").first();
+      expect(this.toggleReactionsLink.text().trim()).toBe("1 reaction");
+      expect(this.toggleReactionsLink.attr("href")).toBe("/posts/" + postGuid + "/comments.mobile");
+    });
+  });
+
+  describe("bottomBarLazy", function(){
+    beforeEach(function() {
+      spec.loadFixture("aspects_index_mobile_post_with_comments");
+      this.bottomBar = $(".bottom-bar").first();
+      this.bottomBarLazy = Diaspora.Mobile.Comments.bottomBarLazy(this.bottomBar);
+    });
+
+    it("shows and hides the loader", function(){
+      expect(this.bottomBarLazy.loader()).toHaveClass("hidden");
+      this.bottomBarLazy.showLoader();
+      expect(this.bottomBarLazy.loader()).not.toHaveClass("hidden");
+      this.bottomBarLazy.hideLoader();
+      expect(this.bottomBarLazy.loader()).toHaveClass("hidden");
+    });
+
+    it("activates the bottom bar", function(){
+      expect(this.bottomBar).toHaveClass("inactive");
+      expect(this.bottomBar).not.toHaveClass("active");
+      expect(this.bottomBarLazy.getShowCommentsLink()).not.toHaveClass("active");
+      expect(this.bottomBarLazy.getShowCommentsLink().find("i")).toHaveClass("entypo-chevron-down");
+      this.bottomBarLazy.activate();
+      expect(this.bottomBar).not.toHaveClass("inactive");
+      expect(this.bottomBar).toHaveClass("active");
+      expect(this.bottomBarLazy.getShowCommentsLink()).toHaveClass("active");
+      expect(this.bottomBarLazy.getShowCommentsLink().find("i")).toHaveClass("entypo-chevron-up");
+    });
+
+    it("deactivates the bottom bar", function(){
+      this.bottomBarLazy.activate();
+      expect(this.bottomBar).not.toHaveClass("inactive");
+      expect(this.bottomBar).toHaveClass("active");
+      expect(this.bottomBarLazy.getShowCommentsLink()).toHaveClass("active");
+      expect(this.bottomBarLazy.getShowCommentsLink().find("i")).toHaveClass("entypo-chevron-up");
+      this.bottomBarLazy.deactivate();
+      expect(this.bottomBar).toHaveClass("inactive");
+      expect(this.bottomBar).not.toHaveClass("active");
+      expect(this.bottomBarLazy.getShowCommentsLink()).not.toHaveClass("active");
+      expect(this.bottomBarLazy.getShowCommentsLink().find("i")).toHaveClass("entypo-chevron-down");
     });
   });
 });